+++ packages/editor/php/StyleDefinitions/MinWidth.php
@@ -3,11 +3,11 @@
 namespace Blockera\Editor\StyleDefinitions;
 
 use Blockera\Editor\StyleDefinitions\Contracts\StandardDefinition;
-use Blockera\Editor\StyleDefinitions\Traits\SimpleDefinitionTrait;
+use Blockera\Editor\StyleDefinitions\Traits\WidthDefinitionTrait;
 
-class MinWidth extends BaseStyleDefinition implements StandardDefinition {
+class MinWidth extends BaseStyleDefinition {
 
-    use SimpleDefinitionTrait;
+    use WidthDefinitionTrait;
 
     public function getCssProperty(): string {